requirements.txt
discord.py
python-dotenv
flask
mysql-connector-python